fixed: CDVDPlayerVideo needs to store the framerate as a double, float doesn't have enough resolution for CVideoReferenceClock
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@35196 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/dvdplayer/DVDPlayerVideo.cpp
@@ -159,7 +159,7 @@ bool CDVDPlayerVideo::OpenStream( CDVDStreamInfo &hint )
 {
 
   if (hint.fpsrate && hint.fpsscale)
-    m_fFrameRate = (float)hint.fpsrate / hint.fpsscale;
+    m_fFrameRate = (double)hint.fpsrate / hint.fpsscale;
   else
     m_fFrameRate = 25;
